Move FeedUIComposer helper classes to separate files
