+++ EssentialFeed/EssentialFeediOS/FeedUI/Composers/FeedUIComposer.swift
@@ -14,7 +14,7 @@ public final class FeedUIComposer {
     public static func feedComposedWith(feedLoader: FeedLoader, imageLoader: FeedImageDataLoader) -> FeedViewController {
         let presentationAdapter = FeedLoaderPresentationAdapter(feedLoader: MainQueueDisptachDecorator(decoratee: feedLoader))
         
-        let feedViewController = FeedViewController.makeWith(delegate: presentationAdapter, title: FeedPresenter.title)
+        let feedViewController = FeedUIComposer.makeWith(delegate: presentationAdapter, title: FeedPresenter.title)
 
         presentationAdapter.presenter = FeedPresenter(
             loadingView: WeakRefVirtualProxy(feedViewController),
@@ -23,10 +23,8 @@ public final class FeedUIComposer {
                 
         return feedViewController
     }
-}
-
-private extension FeedViewController {
-    static func makeWith(delegate: FeedViewControllerDelegate, title: String) -> FeedViewController {
+    
+    private static func makeWith(delegate: FeedViewControllerDelegate, title: String) -> FeedViewController {
         let bundle = Bundle(for: FeedViewController.self)
         let storyboard = UIStoryboard(name: "Feed", bundle: bundle)
         let feedViewController = storyboard.instantiateInitialViewController() as! FeedViewController
@@ -34,106 +32,5 @@ private extension FeedViewController {
         feedViewController.title = title
 
         return feedViewController
-    }
-}
-
-private final class WeakRefVirtualProxy<T: AnyObject> {
-    weak var object: T?
-    
-    init(_ object: T? = nil) {
-        self.object = object
-    }
-}
-
-extension WeakRefVirtualProxy: FeedLoadingView where T: FeedLoadingView {
-    func display(_ viewModel: FeedLoadingViewModel) {
-        object?.display(viewModel)
-    }
-}
-
-extension WeakRefVirtualProxy: FeedImageView where T: FeedImageView, T.Image == UIImage {
-    func display(_ viewModel: FeedImagePresentableModel<UIImage>) {
-        object?.display(viewModel)
-    }
-}
-
-private final class FeedViewAdapter: FeedView {
-    private weak var controller: FeedViewController?
-    private let loader: FeedImageDataLoader
-    
-    init(controller: FeedViewController, loader: FeedImageDataLoader) {
-        self.controller = controller
-        self.loader = loader
-    }
-    
-    func display(_ viewModel: FeedViewModel) {
-        controller?.tableModel = viewModel.feed.map { model in
-            let adapter = FeedImageDataLoaderPresentationAdapter<WeakRefVirtualProxy<FeedImageCellController>, UIImage>(model: model, imageLoader: loader)
-            let view = FeedImageCellController(delegate: adapter)
-            adapter.presenter = FeedImagePresenter(view: WeakRefVirtualProxy(view),
-                                                   imageTransformator: UIImage.init)
-            return view
-        }
-    }
-}
-
-private final class FeedLoaderPresentationAdapter: FeedViewControllerDelegate {
-    private let feedLoader: FeedLoader
-    var presenter: FeedPresenter?
-    
-    init(feedLoader: FeedLoader) {
-        self.feedLoader = feedLoader
-    }
-    
-    func didRequestFeedRefresh() {
-        presenter?.didStartLoadingFeed()
-        
-        feedLoader.load { [weak presenter] result in
-            switch result {
-            case let .success(feed):
-                presenter?.didFinishLoading(with: feed)
-            case let .failure(error):
-                presenter?.didFinishLoading(with: error)
-            }
-        }
-    }
-}
-
-private final class FeedImageDataLoaderPresentationAdapter<View: FeedImageView, Image>: FeedImageCellControllerDelegate where View.Image == Image {
-    private let model: FeedImage
-    private let imageLoader: FeedImageDataLoader
-    private var task: FeedImageDataLoaderTask?
-
-    var presenter: FeedImagePresenter<View, Image>?
-    
-    init(model: FeedImage, imageLoader: FeedImageDataLoader) {
-        self.model = model
-        self.imageLoader = imageLoader
-    }
-
-    func didRequestImage() {
-        presenter?.didStartLoadingImage(for: model)
-        
-        task = self.imageLoader.loadImageData(from: model.url) { [weak self] result in
-            self?.handle(result)
-        }
-    }
-
-    private func handle(_ result: FeedImageDataLoader.Result) {
-        switch result {
-        case let .success(imageData):
-            presenter?.didFinishLoadingImageData(imageData, for: model)
-            
-        case let .failure(error):
-            presenter?.didFinishLoadingImageData(with: error, for: model)
-        }
-    }
-
-    func preload() {
-        didRequestImage()
-    }
-    
-    func didCancelImageRequest() {
-        task?.cancel()
     }
 }
